refactor: webpack 4 backward capability
