+++ package.json
@@ -39,7 +39,7 @@
     "dist"
   ],
   "peerDependencies": {
-    "webpack": "^5.0.0"
+    "webpack": "^3.0.0 || ^4.0.0 || ^5.0.0"
   },
   "dependencies": {
     "@webpack-contrib/schema-utils": "^1.0.0-beta.0",
