+++ src/index.js
@@ -14,10 +14,9 @@ import validateOptions from '@webpack-contrib/schema-utils';
 import NodeTargetPlugin from 'webpack/lib/node/NodeTargetPlugin';
 import SingleEntryPlugin from 'webpack/lib/SingleEntryPlugin';
 import WebWorkerTemplatePlugin from 'webpack/lib/webworker/WebWorkerTemplatePlugin';
-import getLazyHashedEtag from 'webpack/lib/cache/getLazyHashedEtag';
-
-import getWorker from './workers/';
 import WorkerLoaderError from './Error';
+import supportWebpack5 from './supportWebpack5';
+import supportWebpackPrior5 from './supportWebpackPrior5';
 
 export default function loader() {}
 
@@ -71,61 +70,12 @@ export function pitch(request) {
     worker.compiler
   );
 
-  worker.compiler.runAsChild((err, entries, compilation) => {
-    if (err) {
-      return cb(err);
-    }
-
-    if (entries[0]) {
-      worker.file = [...entries[0].files][0];
-
-      const cacheIdent = `${
-        worker.compiler.compilerPath
-      }/worker-loader/${__dirname}/${this.resource}`;
-      const cacheETag = getLazyHashedEtag(compilation.assets[worker.file]);
-
-      return worker.compiler.cache.get(
-        cacheIdent,
-        cacheETag,
-        (err, content) => {
-          if (err) {
-            return cb(err);
-          }
-
-          if (options.fallback === false) {
-            delete this._compilation.assets[worker.file];
-          }
-
-          if (content) {
-            return cb(null, content);
-          }
-
-          worker.factory = getWorker(
-            worker.file,
-            compilation.assets[worker.file].source(),
-            options
-          );
-
-          const newContent = `module.exports = function() {\n  return ${
-            worker.factory
-          };\n};`;
-
-          return worker.compiler.cache.store(
-            cacheIdent,
-            cacheETag,
-            newContent,
-            (err) => {
-              if (err) {
-                return cb(err);
-              }
-
-              return cb(null, newContent);
-            }
-          );
-        }
-      );
-    }
-
-    return cb(null, null);
-  });
+  if (
+    worker.compiler.cache &&
+    typeof worker.compiler.cache.get === 'function'
+  ) {
+    supportWebpack5.call(this, worker, options, cb);
+  } else {
+    supportWebpackPrior5.call(this, worker, request, options, cb);
+  }
 }
