+++ src/supportWebpackPrior5.js
@@ -1,0 +1,48 @@
+import getWorker from './workers';
+
+export default function(worker, request, options, cb) {
+  const subCache = `subcache ${__dirname} ${request}`;
+
+  worker.compilation = (compilation) => {
+    if (compilation.cache) {
+      if (!compilation.cache[subCache]) {
+        compilation.cache[subCache] = {};
+      }
+
+      compilation.cache = compilation.cache[subCache];
+    }
+  };
+
+  if (worker.compiler.hooks) {
+    const plugin = { name: 'WorkerLoader' };
+
+    worker.compiler.hooks.compilation.tap(plugin, worker.compilation);
+  } else {
+    worker.compiler.plugin('compilation', worker.compilation);
+  }
+
+  worker.compiler.runAsChild((err, entries, compilation) => {
+    if (err) return cb(err);
+
+    if (entries[0]) {
+      worker.file = entries[0].files[0];
+
+      worker.factory = getWorker(
+        worker.file,
+        compilation.assets[worker.file].source(),
+        options
+      );
+
+      if (options.fallback === false) {
+        delete this._compilation.assets[worker.file];
+      }
+
+      return cb(
+        null,
+        `module.exports = function() {\n  return ${worker.factory};\n};`
+      );
+    }
+
+    return cb(null, null);
+  });
+}
